move get_metric_fns to utils
